+++ src/clj_kondo/impl/analyzer.clj
@@ -2110,34 +2110,28 @@
     (concat fana
             (analyze-children ctx f-args false))))
 
-(defn analyze-assoc [ctx expr]
-  (let [children (rest (:children expr))
-        [_obj & ks+vs] children
-        ks (take-nth 2 ks+vs)
-        constant-ks (filter (fn [node]
+(defn- analyze-associative [ctx children fn-name ks]
+  (let [constant-ks (filter (fn [node]
                               (or (utils/constant? node)
                                   (utils/symbol-token? node))) ks)]
     (doseq [[k n] (frequencies constant-ks)]
       (when (> n 1)
         (findings/reg-finding! ctx (assoc (meta k)
                                           :filename (:filename ctx)
-                                          :type :duplicate-key-in-assoc
-                                          :message (str "Duplicate key in assoc: " k)))))
-    (analyze-children ctx children false)))
+                                          :type :duplicate-key-in-associative
+                                          :message (str "Duplicate key in " fn-name ": " k))))))
+  (analyze-children ctx children false))
+
+(defn analyze-assoc [ctx expr]
+  (let [[fn-name & children] (:children expr)
+        [_obj & ks+vs] children
+        ks (take-nth 2 ks+vs)]
+    (analyze-associative ctx children fn-name ks)))
 
 (defn analyze-dissoc [ctx expr]
-  (let [children (rest (:children expr))
-        [_obj & ks] children
-        constant-ks (filter (fn [node]
-                              (or (utils/constant? node)
-                                  (utils/symbol-token? node))) ks)]
-    (doseq [[k n] (frequencies constant-ks)]
-      (when (> n 1)
-        (findings/reg-finding! ctx (assoc (meta k)
-                                          :filename (:filename ctx)
-                                          :type :duplicate-key-in-dissoc
-                                          :message (str "Duplicate key in dissoc: " k)))))
-    (analyze-children ctx children false)))
+  (let [[fn-name & children] (:children expr)
+        [_obj & ks] children]
+    (analyze-associative ctx children fn-name ks)))
 
 (defn analyze-ns-unmap [ctx base-lang lang ns-name expr]
   (let [[ns-expr sym-expr :as children] (rest (:children expr))]